removed (Basic)ConllChainNer, replaced with new ConllChainNer that extends ChainNer
